bonus functions enabled(not implemented)
Allows for bonus functions to actually fire when a card wins the rps comparison. This does not, however implement any of these bonus functions, and if you try and start the game in the play button, it will break really bad so don't do that. This makes putting the functions and their appropriate values in the card.txt top priority for friday so the game actually works.